Renamed sklearn to sklearn classifier and reformatted files to pep8
